overhauled the three CI/CD pipelines, making the stricter and supporting three pipelines. NOTE: This commit SHOULD FAIL and will be squashed later
