--- src/tui.rs
@@ -1,1 +1,0 @@
-pub mod app;
