+++ src/ui/ui.rs
@@ -1,0 +1,6 @@
+use ratatui::Frame;
+
+use crate::ui::app::{App, CurrentScreen};
+
+
+pub fn render(frame: &mut Frame, app: &mut App) {}
